[IMP] point_of_sale, **: move fields for category timing
**: pos_self_order

In this commit:
- Moved the fields hour_after and hour_until from the pos_self_order
module to the point_of_sale module in the pos.category model.

task-4319531
