+++ addons/pos_self_order/views/pos_category_views.xml
@@ -5,14 +5,8 @@
         <field name="model">pos.category</field>
         <field name="inherit_id" ref="point_of_sale.product_pos_category_form_view"/>
         <field name="arch" type="xml">
-            <xpath expr="//group[@name='sequence']" position="after">
-                <div class="row col-lg-12">
-                    <span class="text-900 fw-bold">Available between
-                        <span class="col-lg-4 o_light_label">
-                            <a class="o-tooltip"><sup title="Only works for kiosk and mobile">?</sup></a>
-                        </span>
-                        <field name="hour_after" class="oe_inline text-center" widget="float_time"/> and <field name="hour_until" class="oe_inline text-center" widget="float_time"/></span>
-                </div>
+            <xpath expr="//div[@name='available_between']" position="attributes">
+                <attribute name="invisible">0</attribute>
             </xpath>
         </field>
     </record>
